Base controllers, models, repos, services

+++ src/main/java/com/example/ecommerce_api/Controllers/ProductController.java
@@ -1,7 +1,6 @@
 package com.example.ecommerce_api.Controllers;
 
 import com.example.ecommerce_api.Models.Product;
-import com.example.ecommerce_api.Repositories.ProductRepository;
 import com.example.ecommerce_api.Services.ProductService;
 import org.springframework.web.bind.annotation.*;
 
@@ -17,14 +16,13 @@ public class ProductController {
         this.productService = productService;
     }
 
-
     @GetMapping("")
     public List<Product> getAllProducts(){
         return productService.getAllProducts();
     }
 
     @GetMapping("/{id}")
-    public Optional<Product> getProductById(@PathVariable("id") Integer product_id){
+    public Optional<Product> getProductById(@PathVariable("id") Long product_id){
         return productService.getProductById(product_id);
     }
