+++ src/main/java/com/example/ecommerce_api/Repositories/ProductRepository.java
@@ -3,6 +3,8 @@ package com.example.ecommerce_api.Repositories;
 import com.example.ecommerce_api.Models.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductRepository extends JpaRepository<Product, Integer> {
+import java.util.List;
+
+public interface ProductRepository extends JpaRepository<Product, Long> {
 
 }
